refactor: move showNotification tests to their own file
