+++ src/game-logic/reducers/showNotification.test.js
@@ -1,0 +1,20 @@
+import { showNotification } from './showNotification'
+
+describe('showNotification', () => {
+  test('sets notification state', () => {
+    const { latestNotification, todaysNotifications } = showNotification(
+      { showNotifications: true, todaysNotifications: [] },
+      'foo'
+    )
+    const notificationObject = { message: 'foo', severity: 'info' }
+    expect(latestNotification).toEqual(notificationObject)
+    expect(todaysNotifications).toEqual([{ ...notificationObject }])
+  })
+
+  test('does not show redundant notifications', () => {
+    const state = showNotification({ todaysNotifications: [] }, 'foo')
+
+    const { todaysNotifications } = showNotification(state, 'foo')
+    expect(todaysNotifications).toEqual([{ message: 'foo', severity: 'info' }])
+  })
+})
